edit event time
